+++ gmf_core/CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## Unreleased
+
+### Features
+
+- Enhanced memory alignment support for GMF data_bus and payload
+- Add a breaking change for the `esp_gmf_fifo_set_align` api usage
+
 ## v0.8.4
 
 ### Features